+++ agent-akka/src/test/resources/META-INF/monitor/agent.conf
@@ -1,6 +1,6 @@
 org.eigengo.monitor.agent {
     output {
-        class: "org.eigengo.monitor.agent.akka.TestCounterInterface"
+        class: "org.eigengo.monitor.TestCounterInterface"
     }
