+++ agent/src/main/scala/org/eigengo/monitor/agent/CommonAgentConfiguration.scala
