+++ output/pom.xml
@@ -1,8 +1,8 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <artifactId>output-api</artifactId>
+    <artifactId>output</artifactId>
     <packaging>jar</packaging>
-    <name>Output API</name>
+    <name>Output base</name>
 
     <parent>
         <groupId>org.eigengo.monitor</groupId>
