+++ output/src/main/java/org/eigengo/monitor/output/CounterInterface.java
