+++ output/src/main/java/org/eigengo/monitor/output/NullCounterInterface.java
